refactor: extract viz/ subpackage (viz.py + circuit_viz.py)
- src/pdft/viz.py         -> src/pdft/viz/loss.py
- src/pdft/circuit_viz.py -> src/pdft/viz/circuit.py
- tests/test_viz.py       -> tests/viz/test_loss.py

Examples and benchmarks updated to use the new paths via re-exports
in src/pdft/viz/__init__.py. No logic change.
